BUG: Be more careful assigning to avoid FPE overflow

+++ Testing/Code/Numerics/Statistics/itkSampleToHistogramFilterTest7.cxx
@@ -77,18 +77,17 @@ int itkSampleToHistogramFilterTest7( int, char * [] )
   // Populate the Sample
   for( unsigned int i=0; i < histogramSize[0]; i++ )
     {
-    measure[0] = minimum[0] + i;
+    measure[0] = minimum[0] + static_cast<double>(i);
     for( unsigned int j=0; j < histogramSize[1]; j++ )
       {
-      measure[1] = minimum[1] + j;
+      measure[1] = minimum[1] + static_cast<double>(j);
       for( unsigned int k=0; k < histogramSize[2]; k++ )
         {
-        measure[2] = minimum[2] + k;
+        measure[2] = minimum[2] + static_cast<double>(k);
         sample->PushBack( measure );
         }
       }
     }
-
 
   filter->SetInput( sample );
